+++ src/Module.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Hirtz\Media;
 
 use Hirtz\Media\Models\Collections\FolderCollection;
-use Hirtz\Media\Models\interfaces\FileRelationInterface;
+use Hirtz\Media\Models\Interfaces\FileRelationInterface;
 use Hirtz\Skeleton\Filters\PageCache;
 use Hirtz\Skeleton\Modules\ModuleTrait;
 use Yii;