+++ tests/unit/models/traits/MetaImageTraitTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Hirtz\Media\tests\unit\Models\Traits;
 
 use Codeception\Test\Unit;
-use Hirtz\Media\Models\interfaces\AssetInterface;
+use Hirtz\Media\Models\Interfaces\AssetInterface;
 use Hirtz\Media\Models\Traits\MetaImageTrait;
 use Hirtz\Media\tests\data\Models\TestAsset;